+++ src/main/java/org/conjur/jenkins/conjursecrets/ConjurSecretCredentialsBinding.java
@@ -28,6 +28,10 @@ import hudson.model.Run;
 import hudson.model.TaskListener;
 import jenkins.model.Jenkins;
 
+/**
+ * ConjurSecretCredentialsBinding entry level class to when build is invoked to
+ * authorize and retrieve secrets
+ */
 public class ConjurSecretCredentialsBinding extends MultiBinding<ConjurSecretCredentials> {
 
 	@Symbol("conjurSecretCredential")
@@ -56,8 +60,6 @@ public class ConjurSecretCredentialsBinding extends MultiBinding<ConjurSecretCre
 
 	private String credentialsId;
 
-	// private ModelObject context;
-
 	private boolean isParent;
 
 	public boolean isParent() {
@@ -74,9 +76,13 @@ public class ConjurSecretCredentialsBinding extends MultiBinding<ConjurSecretCre
 		this.credentialsId = credentialsId;
 	}
 
+	/**
+	 * Bind method invoked on Jenkins build process
+	 */
 	// @Override
 	public MultiEnvironment bind(Run<?, ?> build, FilePath workSpace, Launcher launcher, TaskListener listener)
 			throws IOException, InterruptedException {
+		long start = System.nanoTime();
 		LOGGER.log(Level.FINE, "**** binding **** : " + build);
 		ConjurCredentialStore store = ConjurCredentialStore.getAllStores()
 				.get(String.valueOf(build.getParent().hashCode()));
@@ -86,11 +92,7 @@ public class ConjurSecretCredentialsBinding extends MultiBinding<ConjurSecretCre
 			store.getProvider().getStore(build);
 		}
 
-		// ConjurSecretCredentials conjurSecretCredential = getCredentials(build);
 		ConjurSecretCredentials conjurSecretCredential = getCredentialsFor(build);
-		// LOGGER.log(Level.FINE, "Context Setin binding class" +
-		// conjurSecretCredential.getDisplayName());
-
 		LOGGER.log(Level.FINE, "Get Parent flage status", isParent);
 		if (!isParent) {
 			LOGGER.log(Level.FINE, "Context Set");
@@ -107,34 +109,17 @@ public class ConjurSecretCredentialsBinding extends MultiBinding<ConjurSecretCre
 				}
 			}
 
-			/*
-			 * Item item = build.getParent();
-			 * 
-			 * if(item !=null) { String parentName =
-			 * ((AbstractItem)(build.getParent()).getParent()).getParent().getDisplayName();
-			 * LOGGER.log(Level.FINE, "Context Set not for parent 1"+parentName);
-			 * if(!parentName.isEmpty() && !parentName.equalsIgnoreCase("Jenkins")) {
-			 * LOGGER.log(Level.FINE,
-			 * "Context Set not for parent 2"+build.getParent().getDisplayName());
-			 * 
-			 * conjurSecretCredential.setContext(((AbstractItem)(build.getParent()).
-			 * getParent()).getParent());
-			 * 
-			 * 
-			 * 
-			 * } else { LOGGER.log(Level.FINE,
-			 * "Context Set not for parent 3"+build.getParent().getParent().getDisplayName()
-			 * ); conjurSecretCredential.setContext(build.getParent().getParent()); } }
-			 */
-
 		}
-
+		long end = System.nanoTime();
+		long execution = end - start;
+	    LOGGER.log(Level.OFF,"Execution of Class ConjurSecretCredentialsBinding -->Method bind() time: "+ execution/1000000d + " milliseconds");
 		return new MultiEnvironment(
 				Collections.singletonMap(variable, conjurSecretCredential.getSecret().getPlainText()));
 	}
 
-	private final @Nonnull <C> C getCredentialsFor(@Nonnull Run<?, ?> build) throws IOException {
 
+	private final @Nonnull <C> C getCredentialsFor(@Nonnull Run<?, ?> build) throws IOException ,InterruptedException{
+		long start = System.nanoTime();
 		IdCredentials cred = CredentialsProvider.findCredentialById(credentialsId, IdCredentials.class, build);
 		LOGGER.log(Level.FINE, "Calling getCredential For1" + build.getFullDisplayName());
 		String newCredentialId = "";
@@ -143,7 +128,7 @@ public class ConjurSecretCredentialsBinding extends MultiBinding<ConjurSecretCre
 
 			setParent(true);
 
-			Item item = (Item) build.getParent(); // Item item =(Item) build;
+			Item item = (Item) build.getParent(); 
 
 			if (item != null) {
 				LOGGER.log(Level.FINE, "Item Name" + item.getParent().getDisplayName());
@@ -151,18 +136,17 @@ public class ConjurSecretCredentialsBinding extends MultiBinding<ConjurSecretCre
 				LOGGER.log(Level.FINE, "CredentialId after removing ${}" + newCredentialId);
 
 				ConjurSecretCredentials conjurSecretCredential = null;
-
-				// conjurSecretCredential =
-				// ConjurSecretCredentials.credentialWithID(newCredentialId, item);
-				conjurSecretCredential = ConjurSecretCredentials.credentialWithID(newCredentialId, build);
-				LOGGER.log(Level.FINE, "From Binding Credential" + conjurSecretCredential.getDisplayName());
+				
+				conjurSecretCredential = ConjurSecretCredentials.credentialWithID(newCredentialId, item);
+				LOGGER.log(Level.FINE, "From Binding Credential" + conjurSecretCredential);
 
 				cred = conjurSecretCredential;
+				if(cred==null)
+				{
+					throw new CredentialNotFoundException("Could not find credentials entry with ID '" + credentialsId + "'");
+				}
+				
 
-				// CredentialsProvider.findCredentialById(credentialsId,
-				// IdCredentials.class,build);
-				// throw new CredentialNotFoundException("Could not find credentials entry with
-				// // ID '" + credentialsId + "'");
 			}
 		}
 
@@ -172,16 +156,21 @@ public class ConjurSecretCredentialsBinding extends MultiBinding<ConjurSecretCre
 		}
 
 		Descriptor<?> expected = Jenkins.getActiveInstance().getDescriptor(type());
+		long end = System.nanoTime();
+		long execution = end - start;
+	    LOGGER.log(Level.OFF,"Execution of Class ConjurSecretCredentialsBinding -->Method getCredentialsFor() time: "+ execution/1000000d + " milliseconds");
 		throw new CredentialNotFoundException(
-				"Credentials '" + credentialsId + "' is of type '" + cred.getDescriptor().getDisplayName() + "' where '"
+				"Credentials '" + credentialsId + "' not found '" + cred + "' where '"
 						+ (expected != null ? expected.getDisplayName() : type().getName()) + "' was expected");
 
 	}
 
+	/** @return variable */
 	public String getVariable() {
 		return this.variable;
 	}
 
+	/** set the variable */
 	@DataBoundSetter
 	public void setVariable(String variable) {
 		LOGGER.log(Level.FINE, "Setting variable to {0}", variable);